# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Loyalty Engine API
  version: 1.0.0
  description: |
    Multi-tenant B2B loyalty logic engine with API-key secured endpoints.
servers:
  - url: http://localhost:3000
    description: Local development
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    UpdateRatioRequest:
      type: object
      properties:
        pointsRatio:
          type: number
          exclusiveMinimum: 0
      required: [pointsRatio]
    UpdateRatioResponse:
      type: object
      properties:
        organizationId:
          type: string
          format: uuid
        pointsRatio:
          type: number
        updatedAt:
          type: string
          format: date-time
      required: [organizationId, pointsRatio, updatedAt]
...
paths:
...
  /api/v1/organizations/{organizationId}/ratio:
    patch:
      tags: [Organizations]
      summary: Update points ratio
      operationId: updatePointsRatio
      security:
        - ApiKeyAuth: []
      parameters:
        - name: organizationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateRatioRequest'
            example:
              pointsRatio: 12
      responses:
        '200':
          description: Ratio updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpdateRatioResponse'
        '400':
          description: Validation or processing error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...